pymongo: move old trace_mongo_client to patch.py
It seems more logical to put it in the patch file rather than in the
client file
