+++ ddtrace/contrib/pymongo/__init__.py
@@ -31,6 +31,5 @@ required_modules = ['pymongo']
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .client import trace_mongo_client
-        from .patch import patch
+        from .patch import trace_mongo_client, patch
         __all__ = ['trace_mongo_client', 'patch']
