+++ ddtrace/contrib/pymongo/client.py
@@ -13,21 +13,12 @@ from ...constants import ANALYTICS_SAMPLE_RATE_KEY, SPAN_MEASURED_KEY
 from ...ext import SpanTypes, mongo as mongox, net as netx
 from ...internal.logger import get_logger
 from ...settings import config
-from ...utils.deprecation import deprecated
 from .parse import parse_spec, parse_query, parse_msg
 
 # Original Client class
 _MongoClient = pymongo.MongoClient
 
 log = get_logger(__name__)
-
-
-@deprecated(message='Use patching instead (see the docs).', version='1.0.0')
-def trace_mongo_client(client, tracer, service=mongox.SERVICE):
-    traced_client = TracedMongoClient(client)
-    ddtrace.Pin(service=service, tracer=tracer).onto(traced_client)
-    return traced_client
-
 
 class TracedMongoClient(ObjectProxy):
 
@@ -245,7 +236,6 @@ class TracedSocket(ObjectProxy):
         if self.address:
             _set_address_tags(s, self.address)
         return s
-
 
 def normalize_filter(f=None):
     if f is None:
